types: added assignment operators to RTT::types::carray to assign from boost::serialization::array and boost::array
The `boost::array` operator is required to assign an array data source from a `boost::array`, e.g. a fixed-sized
field of a ROS message. The `boost::serialization::array` assignment operator was added for completeness as carray also
has a `boost::serialization::array` constructor.

The value type of the RHS of the assignment may differ from the value type of the carray as long as the elements are assignable.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>