Cleanup of CMAKE_**_ADD flags.
See: https://github.com/orocos-toolchain/rtt/pull/88

This causes a set of COMPILE_OPTIONS with CXX flags.
This causes that tlsf.c got compiled with cxx options.
Which is failing with cxx11.

The env settings of CMAKE_XXX_FLAGS is applied by default to all
packages. There is no reason to add it to custom variables.