internal: return WriteFailure instead of NotConnected from ConnOutputEndPoint::write()
When a connection is added to an OutputPort and the init flag is set to true in the connection policy,
the port invokes write() on the new channel element to be connected. For WriteShared connections, this channel
element is directly the ConnOutputEndPoint of the InputPort to be connected, which cannot be written. In that
case, the write() call should simply fail (return WriteFailure), but not abort the connection process by
returning NotConnected.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>