Moved portqueue from class TaskContext to the ExecutionEngine
This is an improved version of f26633764e3cff394206be5cd627ec795291d80c, which solves
the problem that operation calls and port callbacks could be interweaved unexpectedly.

A new method dataOnPortCallback(PortInterface*) has been added to the TaskContext
interface which can be overridden in subclasses and which will be called exactly once
per incoming sample on an EventPort by the ExecutionEngine.

The former internal methods TaskContext::dataOnPortCallback() and TaskContext::dataOnPortRemoved()
methods for callback management have been renamed to setDataOnPortCallback() and
removeDataOnPortCallback() for clarity.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>