engine: fixed forwarding of messages and port callbacks received in slaves to the master engine
This commit fixes a bug introduced in 0f71761 that broke port callbacks in components running with a SlaveActivity.
All message events and port callbacks must be forwarded to the master's ExecutionEngine and whenever someone needs
to block until an event has been processed we need to wait on the master's condition variable.
waitForMessagesInternal() or waitAndProcessMessages() must never be called on slave engines.

Furthermore this PR blocks messages and port callbacks of slave engines if the component is in the FatalError state,
for consistency with non-slave components.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>